[refactor] #48 presentation layer  ui/main package 생성 mainScreen 책임분리
