+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:theme="@style/Theme.BBANGZIPANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.ui.navigator.MainActivity"
+            android:name=".presentation.ui.main.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.BBANGZIPANDROID">
             <intent-filter>
