+++ app/src/main/java/org/android/bbangzip/presentation/common/type/BreadType.kt
@@ -1,4 +1,4 @@
-package org.android.bbangzip.presentation.type
+package org.android.bbangzip.presentation.common.type
 
 import androidx.annotation.DrawableRes
 import org.android.bbangzip.R
