+++ app/src/main/java/org/android/bbangzip/presentation/common/type/TimerViewType.kt
@@ -1,4 +1,4 @@
-package org.android.bbangzip.presentation.type
+package org.android.bbangzip.presentation.common.type
 
 import androidx.annotation.StringRes
 import androidx.compose.runtime.Stable
