+++ app/src/main/java/org/android/bbangzip/presentation/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package org.android.bbangzip.presentation.ui.navigator
+package org.android.bbangzip.presentation.ui.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
