+++ app/src/main/java/org/android/bbangzip/presentation/ui/main/MainScreen.kt
@@ -1,4 +1,4 @@
-package org.android.bbangzip.presentation.ui.navigator
+package org.android.bbangzip.presentation.ui.main
 
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.asPaddingValues
@@ -12,8 +12,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import okhttp3.internal.toImmutableList
-import org.android.bbangzip.presentation.type.BottomNavigationType
-import org.android.bbangzip.presentation.ui.navigator.component.BottomNavigationBar
+import org.android.bbangzip.presentation.ui.main.BottomNavigationType
+import org.android.bbangzip.presentation.ui.main.component.BottomNavigationBar
+import org.android.bbangzip.presentation.ui.main.component.MainNavHost
 import org.android.bbangzip.presentation.ui.shared.SharedViewModel
 import org.android.bbangzip.ui.theme.BBANGZIPANDROIDTheme
