+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/todo/TimerTodoScreen.kt
@@ -33,7 +33,7 @@ import org.android.bbangzip.presentation.common.component.topbar.BbangZipBaseTop
 import org.android.bbangzip.presentation.common.model.Category
 import org.android.bbangzip.presentation.common.model.ListItem
 import org.android.bbangzip.presentation.common.model.Todo
-import org.android.bbangzip.presentation.type.CategoryColor
+import org.android.bbangzip.presentation.common.type.CategoryColor
 import org.android.bbangzip.presentation.common.util.extension.Gap
 import org.android.bbangzip.ui.theme.BBANGZIPANDROIDTheme
 import org.android.bbangzip.ui.theme.BbangZipTheme
